==> R/00_all.qmd <==
---
title: "00_all"
author: "group22"
format:
  html:
    embed-resources: true
editor: visual
---

{{< include 01_load.qmd >}}

{{< include 02_clean.qmd >}}

{{< include 03_augment.qmd >}}

{{< include 04_describe.qmd >}}

{{< include 05_analysis_1.qmd >}}

{{< include 06_analysis_2.qmd >}}
