==> R/01_load.qmd <==
---
title: "01_load"
author: "group22"
format:
  html:
    embed-resources: true
editor: visual
---

## Load Libraries

```{r message = FALSE}
library("tidyverse")
library("ggridges")
library("RColorBrewer")
```

## Load Data

Loading the data from .csv file that has been downloaded.

```{r message = FALSE}
df <- read_csv("../_raw/project_data.csv")
```

## Save loaded data

Saving the dataset as .tsv file in data directory. Also reading the saved file to make sure it works.

```{r message = FALSE}
write_tsv(df, "../data/01_dat_load.tsv")

df <- read_tsv("../data/01_dat_load.tsv")
```
