==> R/02_clean.qmd <==
---
title: "02_clean"
author: "group22"
format:
  html:
    embed-resources: true
editor: visual
---

```{r echo = FALSE, message = FALSE}
library("tidyverse")
```

## Checking for missing values

In the first part of data cleaning, we focus on identifying and removing any missing values in our dataset. Missing data can significantly impact the quality of our analysis, leading to biased or inaccurate results.

```{r message = FALSE}
df <- read_tsv("../data/01_dat_load.tsv")

df_cleaned <- df |> 
  drop_na()

rows_removed <- nrow(df) - nrow(df_cleaned)
cat("Rows removed because of missing values:", rows_removed, "\n")
```

From the first part we can see that the dataset didn't have any missing variables. If there would have been any missing values, we would have removed those rows with missing values.

## Checking for right type and values of each variable

In the second part we want to ensure that each of the variables in the dataset has expected type. By doing this we can make sure that we don't have any faulty variables and we understand our variables for further analysis.

```{r}
column_types <- df_cleaned |> 
  summarise(across(everything(), class))

print(column_types)
```

We can see that all of our variables are numeric.

## Filtering incorrect values

In the third part of cleaning we want to filter out incorrect values. With the dataset received we also received range of values each numeric variable. So this part is all about making sure that each of the value falls into correct range of values.

```{r}
original_df <- df_cleaned

df_cleaned <- df_cleaned |>
  filter(Diabetes_binary %in% c(0, 1, 2),
         HighBP %in% c(0, 1),
         HighChol %in% c(0, 1),
         CholCheck %in% c(0, 1),
         BMI > 0,
         Smoker %in% c(0, 1),
         Stroke %in% c(0, 1),
         HeartDiseaseorAttack %in% c(0, 1),
         PhysActivity %in% c(0, 1),
         Fruits %in% c(0, 1),
         Veggies %in% c(0, 1),
         HvyAlcoholConsump %in% c(0, 1),
         AnyHealthcare %in% c(0, 1),
         NoDocbcCost %in% c(0, 1),
         GenHlth %in% 1:5,
         MentHlth %in% 0:30,
         PhysHlth %in% 0:30,
         DiffWalk %in% c(0, 1),
         Sex %in% c(0, 1),
         Age %in% 1:13,
         Education %in% 1:6,
         Income %in% 1:8)

rows_removed_after_filter <- nrow(original_df) - nrow(df_cleaned)
cat("Rows removed because of incorrect values:", rows_removed_after_filter, "\n")
```

From this part we can see that the dataset didn't have any out of range values for all of the variables. If there would have been any incorrect values, we would have removed those rows with incorrect values.

## Writing the file

```{r}
write_tsv(df, "../data/02_dat_clean.tsv")
```
