==> R/04_describe.qmd <==
---
title: "04_describe"
author: "group22"
format:
  html:
    embed-resources: true
editor: visual
---

```{r echo = FALSE, message = FALSE}
library("tidyverse")
library("ggridges")
library("RColorBrewer")
```

## Load data

```{r message = FALSE}
df_aug <- read_tsv("../data/03_dat_aug.tsv")
```

## Visualization

1.  **The distribution of BMI in the different age groups, including a comparison between smokers and non-smokers.**

```{r}
plot_1 <- 
  df_aug |> 
  ggplot(aes(x = BMI, y = Age_Range)) +
  geom_density_ridges(aes(fill = Age_Range) ,alpha=0.6) + 
  facet_wrap(~Smoking_Status)+
  labs(x = "BMI",
       y = "Age Group") +
  theme_minimal(base_size = 14)+
  theme(legend.position = "none",
        plot.background = element_rect(fill = "white", color="white")) +
  scale_fill_viridis_d(option = "rocket") +
  labs(title = "BMI distribution over the different age groups", 
       subtitle = "A comparison between smokers and non-smokers")

plot_1
```

As it is illustrated from the plot above, smoking status affects mostly the youngest age groups, namely 18-24 and 25-34, where it is shown that smokers have a generally higher BMI than the non-smokers. For the other age groups, the behavior is almost the same between smokers and non-smokers, so smoking doesn't appear as a main factor influencing BMI for individuals over 35. However, the general tendency is that BMI increases over the age, as it can be seen by the peak.

2.  **Comparison of BMI between income classes, divided into individuals with and without physical activity.**

```{r}
df_aug |> 
  ggplot(mapping = aes(x = BMI,
                       y = fct_relevel(Income_Class,"Poor","Average","Wealthy"))) +
  geom_boxplot(aes(fill = Physically_Active),
               outlier.shape = NA,,alpha=0.6) +
  xlim(0,60) +
  xlab("BMI") +
  ylab("Income Class") +
  theme(legend.position = "none") +
  scale_fill_manual(values = c("aquamarine", "blue")) +
  labs(title = "Comparison of BMI between different income classes", 
       subtitle = "Red: No Physical Activity, Blue: Physical Activity")
```

The graph displays a general decrease of BMI in physically active individuals, comparing to inactive ones. Also, it is evident that BMI gradually increases from the wealthiest to the poorest income group, regardless of the physical activity status.

3.  **Distribution of habits among individuals between genders**

```{r}
plot_2 <- 
  df_aug |> 
  ggplot(mapping = aes(x = fct_relevel(Habits,"Unhealthy","Average","Healthy"),
                       fill = Sex_character)) +
  geom_bar(position = position_dodge(),alpha=0.8) +
  facet_wrap(~Diabetes_Status) +
  theme_minimal(base_size = 14)+
  scale_fill_manual(name = "Gender", values = c("midnightblue", "mediumorchid2")) +
  #scale_fill_manual(values = c("antiquewhite4", "coral2"))
  #scale_fill_viridis_d() +
  theme(legend.position = "right",
        plot.background = element_rect(fill = "white", color="white")) +
  ylab("Count") +
  xlab("Habits") +
  labs(title = "The distribution of habits among individuals", 
       subtitle = "Comparison between 2 genders")

plot_2
```

Among the individuals with healthy habits, it is illustrated that, in both diabetic and non-diabetic cases, more women have healthier habits, whereas in the "Average" habits group, the results are more conflicting. It can be seen that while in the non-diabetic cases, more women have an "Average" lifestyle, it is the opposite for the diabetic counterparts. It is also remarkable that female individuals in the non-diabetic group doubled from the "Average" to the "Healthy" group, while the increase is slighter in the diabetic cases.

On the contrary, regarding the male counterparts, among the diabetic cases, there is not an important change between "Average" and "Healthy" groups, while in the non-diabetic cases, the "Healthy" cases where nearly 2000 more than the "Average" ones.

# Saving Plots

```{r message = FALSE}
plot_1 |> 
  ggsave(filename = "04_key_plot_1.png", path = "../results/")

plot_2 |> 
  ggsave(filename = "04_key_plot_2.png", path = "../results/")
```
